Cedric login

+++ screens/TripsScreen.js
@@ -6,18 +6,15 @@ import {
   Pressable,
   ScrollView,
   SafeAreaView,
-  RefreshControl,
   Dimensions,
   Modal,
 } from "react-native";
 import FontAwesome from "react-native-vector-icons/FontAwesome";
-import React, { useState, useEffect, useCallback } from "react";
+import React, { useState, useEffect } from "react";
 import { useIsFocused } from "@react-navigation/native";
-import { createTripCard } from "../reducers/trips";
 import { updateRefresh } from "../reducers/user";
 import { useDispatch, useSelector } from "react-redux";
 import Footer from "../components/Footer";
-import Constants from "expo-constants";
 import {
   emptySizes,
   updateDayPlan,
@@ -26,31 +23,17 @@ import {
   getTripId,
 } from "../reducers/activ";
 
-const backend = Constants.expoConfig.hostUri.split(`:`)[0];
-
 export default function TripsScreen({ navigation }) {
   const dispatch = useDispatch();
   const isFocused = useIsFocused();
 
   const { token, refresh } = useSelector((state) => state.user.value);
-  const tripCard = useSelector((state) => state.trips.cityCard);
-  const myTrips = useSelector((state) => state.trips.value);
-  const allSizes = useSelector((state) => state.activ.sizesArray);
 
   const [allTrips, setAllTrips] = useState([]);
   const [trigger, setTrigger] = useState(false);
 
-  const [refreshing, setRefreshing] = useState(false);
   const [deleteModalVisible, setDeleteModalVisible] = useState(false);
-  const [dataModalVisible, setDataModalVisible] = useState(false);
   const [id, setId] = useState("");
-
-  const onRefresh = useCallback(() => {
-    setRefreshing(true);
-    setTimeout(() => {
-      setRefreshing(false);
-    }, 2000);
-  }, []);
 
   const emptySizesArray = () => {
     dispatch(emptySizes());
@@ -58,33 +41,34 @@ export default function TripsScreen({ navigation }) {
 
   useEffect(() => {
     if (refresh > 0) {
-      (async () => {
-        const response = await fetch(`https://tripperz-backend.vercel.app/trips/${token}`);
-        const data = await response.json();
-        if (data.trips.length > 0) {
-          setAllTrips(data.trips);
-        } else {
-          setAllTrips([]);
+      (async() => {
+        if (token !== null) {
+          const response = await fetch(`https://tripperz-backend.vercel.app/trips/${token}`);
+          const data = await response.json();
+          if (data.trips.length > 0) {
+            setAllTrips(data.trips);
+          } else {
+            setAllTrips([]);
+          }
         }
       })();
     }
   }, [trigger]);
 
   const handleDelete = async (id) => {
-    console.log("id", id);
-    const response = await fetch(
+    if (token !== null) {
+      const response = await fetch(
       `https://tripperz-backend.vercel.app/trips/${id}/${token}`,
       {
         method: "DELETE",
         headers: { "Content-Type": "application/json" },
-      }
+        }
     );
-    const data = await response.json();
-    if (data.result) {
+      const data = await response.json();
+      if (data.result) {
       setDeleteModalVisible(!deleteModalVisible);
-      setTrigger(!trigger);
-    } else {
-      console.log(data.error);
+        setTrigger(!trigger);
+      }
     }
   };
 
@@ -152,18 +136,19 @@ export default function TripsScreen({ navigation }) {
     return <View />;
   } else {
     if (refresh === 0) {
-      console.log("focus");
       emptySizesArray();
-      (async () => {
-        const response = await fetch(`https://tripperz-backend.vercel.app/trips/${token}`);
-        const data = await response.json();
-        if (data.trips.length > 0) {
-          setAllTrips(data.trips);
-        } else {
-          setAllTrips([]);
-        }
-      })();
-      dispatch(updateRefresh(1));
+      if (token !== null) {
+        (async() => {
+          const response = await fetch(`https://tripperz-backend.vercel.app/trips/${token}`);
+          const data = await response.json();
+          if (data.trips.length > 0) {
+            setAllTrips(data.trips);
+          } else {
+            setAllTrips([]);
+          }
+        })();
+        dispatch(updateRefresh(1));
+      }
     }
   }
